Fold 'multipart' into MultipartRequestStream
