+++ httpx/streams.py
@@ -1,8 +1,13 @@
+import binascii
+import mimetypes
+import os
 import typing
+from io import BytesIO
 from json import dumps as json_dumps
+from pathlib import Path
 from urllib.parse import urlencode
 
-from .multipart import multipart_encode
+from .utils import format_form_param
 
 RequestData = typing.Union[dict, str, bytes, typing.AsyncIterator[bytes]]
 
@@ -127,8 +132,111 @@ class MultipartRequestStream(RequestStream):
     Request content as multipart encoded form data.
     """
 
+    class DataField:
+        """
+        A single form field item, within a multipart form field.
+        """
+
+        def __init__(self, name: str, value: typing.Union[str, bytes]) -> None:
+            if not isinstance(name, str):
+                raise TypeError("Invalid type for name. Expected str.")
+            if not isinstance(value, (str, bytes)):
+                raise TypeError("Invalid type for value. Expected str or bytes.")
+            self.name = name
+            self.value = value
+
+        def render_headers(self) -> bytes:
+            name = format_form_param("name", self.name)
+            return b"".join([b"Content-Disposition: form-data; ", name, b"\r\n\r\n"])
+
+        def render_data(self) -> bytes:
+            return (
+                self.value
+                if isinstance(self.value, bytes)
+                else self.value.encode("utf-8")
+            )
+
+    class FileField:
+        """
+        A single file field item, within a multipart form field.
+        """
+
+        def __init__(
+            self, name: str, value: typing.Union[typing.IO[typing.AnyStr], tuple]
+        ) -> None:
+            self.name = name
+            if not isinstance(value, tuple):
+                self.filename = Path(str(getattr(value, "name", "upload"))).name
+                self.file = (
+                    value
+                )  # type: typing.Union[typing.IO[str], typing.IO[bytes]]
+                self.content_type = self.guess_content_type()
+            else:
+                self.filename = value[0]
+                self.file = value[1]
+                self.content_type = (
+                    value[2] if len(value) > 2 else self.guess_content_type()
+                )
+
+        def guess_content_type(self) -> typing.Optional[str]:
+            if self.filename:
+                return (
+                    mimetypes.guess_type(self.filename)[0] or "application/octet-stream"
+                )
+            else:
+                return None
+
+        def render_headers(self) -> bytes:
+            parts = [
+                b"Content-Disposition: form-data; ",
+                format_form_param("name", self.name),
+            ]
+            if self.filename:
+                filename = format_form_param("filename", self.filename)
+                parts.extend([b"; ", filename])
+            if self.content_type is not None:
+                content_type = self.content_type.encode()
+                parts.extend([b"\r\nContent-Type: ", content_type])
+            parts.append(b"\r\n\r\n")
+            return b"".join(parts)
+
+        def render_data(self) -> bytes:
+            if isinstance(self.file, str):
+                content = self.file
+            else:
+                content = self.file.read()
+            return content.encode("utf-8") if isinstance(content, str) else content
+
     def __init__(self, data: dict, files: dict, boundary: bytes = None) -> None:
-        self.body, self.content_type = multipart_encode(data, files, boundary)
+        body = BytesIO()
+        if boundary is None:
+            boundary = binascii.hexlify(os.urandom(16))
+
+        for field in self.iter_fields(data, files):
+            body.write(b"--%s\r\n" % boundary)
+            body.write(field.render_headers())
+            body.write(field.render_data())
+            body.write(b"\r\n")
+
+        body.write(b"--%s--\r\n" % boundary)
+
+        self.content_type = "multipart/form-data; boundary=%s" % boundary.decode(
+            "ascii"
+        )
+        self.body = body.getvalue()
+
+    def iter_fields(
+        self, data: dict, files: dict
+    ) -> typing.Iterator[typing.Union["FileField", "DataField"]]:
+        for name, value in data.items():
+            if isinstance(value, (list, dict)):
+                for item in value:
+                    yield self.DataField(name=name, value=item)
+            else:
+                yield self.DataField(name=name, value=value)
+
+        for name, value in files.items():
+            yield self.FileField(name=name, value=value)
 
     def get_headers(self) -> typing.Dict[str, str]:
         content_length = str(len(self.body))
